MDXMini, PMDMini: use nonprefixed functions
Let the Player classes extend the modules so that the functions are
available nonprefixed. Also follow the UADE style of calling wrapped
C functions - non-namespced, and always called with parens, to make it
clear when something is a C function versus when it's native Ruby code.

e.g. it looks better and is clearer to do:

mdx_open(mini, filename, dirname)

Than:

MDXMini.mdx_open mini, filename, dirname